Mark as stale only enhancement PRs
If for some reason we miss triaging a PR, the Stale bot won't automatically mark it stale. I think this is a Good Thing (tm), since it doesn't penalize the PR author if we're not diligent with our PR housekeeping.

Issues will no longer be marked stale, since they will be either be kept in Issues as bugs, or converted as discussions.